fixed indexing in parse_table function; added argparse for argument parsing
parse_table originally iterated from j[1:-1] in order to account for
quotation marks which were removed in a later update. Line now reads
row.append(j)

added argparse to parse the input to our script. It allows more
flexible argument input to the script (e.g. any order, marking input as
optional or required, etc.) previous version prompted for password but
then project_name and other options were not longer able to be entered
easily. This update fixes that problem. See script for detailed
comments on argparse.

NOTE: as far as I can tell everything is currently functional except
specifying a subset of participants for downloading (need to consult
with Paul on format of the match). However I have not specifically
tested the -all functionality yet so I do not know if that is working.